server/requirements.txt
PyYAML
pika
psycopg2-binary